WI #2707 Move method to LoggingSystemExtensions
