+++ TypeCobol/Compiler/CupParser/NodeBuilder/ProgramClassBuilder.cs
@@ -1,5 +1,4 @@
-﻿using System.Text;
-using JetBrains.Annotations;
+﻿using JetBrains.Annotations;
 using TypeCobol.Compiler.CodeElements;
 using TypeCobol.Compiler.CodeModel;
 using TypeCobol.Compiler.Nodes;
@@ -15,85 +14,15 @@ namespace TypeCobol.Compiler.CupParser.NodeBuilder
     /// </summary>
     public class ProgramClassBuilder : IProgramClassBuilder
     {
-        private static readonly HashSet<string> _SourceCodeDumpReasons = new HashSet<string>();
-
         internal void LogAnomalousLineIndex() => LogErrorWithOptionalFullSourceCode(nameof(LogAnomalousLineIndex), "Anomalous line index found !");
 
         private void LogInvalidAst(string message) => LogErrorWithOptionalFullSourceCode(nameof(LogInvalidAst), message);
 
-        private void LogErrorWithOptionalFullSourceCode(string reason, string message, bool fullDump = false)
+        private void LogErrorWithOptionalFullSourceCode(string reason, string message, bool includeSourceCode = false, bool includeHistory = false)
         {
-            // Fail immediately in debug. There is no point in dumping source code and this allows to see the error.
-            System.Diagnostics.Debug.Fail(message);
-
-            if (!_SourceCodeDumpReasons.Add(reason)) return; // Dump only once for this reason to avoid huge logs
-
-            // Build a correlation id to group traces
-            string correlationId = DateTime.Now.ToString("s") + "@" + Environment.MachineName;
-            var currentTrace = new StringBuilder();
-
-            if (fullDump)
-            {
-                int traceId = 0;
-
-                // Dump file structure
-                IterateCodeElementsLines(DumpCodeElementTypes);
-
-                // Dump source code
-                IterateCodeElementsLines(codeElementsLine => currentTrace.AppendLine(codeElementsLine.Text));
-
-                // Dump last incremental changes
-                int changeId = -_history.Depth + 1; //Number last changes from -N+1 to 0
-                foreach (var textChangedEvent in _history.TextChangedEvents)
-                {
-                    currentTrace.AppendLine($"change {changeId}:");
-                    foreach (var documentChange in textChangedEvent.DocumentChanges)
-                    {
-                        string text = documentChange.NewLine == null
-                            ? string.Empty
-                            : '"' + documentChange.NewLine.Text + '"';
-                        currentTrace.AppendLine($"{documentChange.Type}@{documentChange.LineIndex} {text}");
-                    }
-
-                    changeId++;
-                }
-
-                void IterateCodeElementsLines(Action<CodeElementsLine> appendLine)
-                {
-                    const int MAX_LINES_PER_LOG = 1000;
-                    for (int i = 0; i < _codeElementsLines.Count; i++)
-                    {
-                        appendLine(_codeElementsLines[i]);
-
-                        if ((i + 1) % MAX_LINES_PER_LOG == 0 || i == _codeElementsLines.Count - 1)
-                        {
-                            var contextData = new Dictionary<string, object>()
-                            {
-                                { $"{correlationId} - part {traceId}", currentTrace.ToString() }
-                            };
-                            LoggingSystem.LogMessage(LogLevel.Error, message, contextData);
-                            currentTrace.Clear();
-                            traceId++;
-                        }
-                    }
-                }
-
-                void DumpCodeElementTypes(CodeElementsLine codeElementsLine)
-                {
-                    if (!codeElementsLine.HasCodeElements) return;
-
-                    foreach (var codeElement in codeElementsLine.CodeElements)
-                    {
-                        currentTrace.AppendLine($"{codeElement.Line}: {codeElement.Type}");
-                    }
-                }
-            }
-            else
-            {
-                currentTrace.Append("No source code dump");
-            }
-
-            LoggingSystem.LogMessage(LogLevel.Error, message, new Dictionary<string, object>() { { $"{correlationId} - changes", currentTrace.ToString() } });
+            var codeElementsLines = includeSourceCode ? _codeElementsLines : null;
+            var history = includeHistory ? _history : null;
+            LoggingSystemExtensions.LogErrorWithSourceCode(reason, message, codeElementsLines, history);
         }
 
         public SyntaxTree SyntaxTree { get; set; }
